bite the bullet and refactor to use `commands/utils/` and `quickpicks/utils/` subdirs
