+++ src/commands/topics.ts
@@ -4,7 +4,6 @@ import { registerCommandWithLogging } from ".";
 import {
   ProduceRecordRequest,
   ProduceRequest,
-  ProduceRequestData,
   ProduceRequestHeader,
   ProduceResponse,
   RecordsV3Api,
@@ -17,28 +16,20 @@ import {
 } from "../clients/sidecar";
 import { MessageViewerConfig } from "../consume";
 import { MESSAGE_URI_SCHEME } from "../documentProviders/message";
-import { DEFAULT_ERROR_NOTIFICATION_BUTTONS, showErrorNotificationWithButtons } from "../errors";
-import { ResourceLoader } from "../loaders";
+import { showErrorNotificationWithButtons } from "../errors";
 import { Logger } from "../logging";
 import { KafkaCluster } from "../models/kafkaCluster";
 import { isCCloud } from "../models/resource";
-import { Schema, Subject } from "../models/schema";
-import { SchemaRegistry } from "../models/schemaRegistry";
+import { Schema } from "../models/schema";
 import { KafkaTopic } from "../models/topic";
-import { ALLOW_OLDER_SCHEMA_VERSIONS, USE_TOPIC_NAME_STRATEGY } from "../preferences/constants";
-import {
-  schemaKindMultiSelect,
-  SchemaKindSelection,
-  schemaSubjectQuickPick,
-  schemaVersionQuickPick,
-  subjectNameStrategyQuickPick,
-} from "../quickpicks/schemas";
+import { schemaKindMultiSelect, SchemaKindSelection } from "../quickpicks/schemas";
 import { loadDocumentContent, LoadedDocumentContent, uriQuickpick } from "../quickpicks/uris";
+import { promptForSchema } from "../quickpicks/utils/schemas";
+import { getSubjectNameStrategy } from "../quickpicks/utils/schemaSubjects";
 import { JSON_DIAGNOSTIC_COLLECTION } from "../schemas/diagnosticCollection";
 import {
   PRODUCE_MESSAGE_SCHEMA,
   ProduceMessage,
-  SchemaInfo,
   SubjectNameStrategy,
 } from "../schemas/produceMessageSchema";
 import { validateDocument } from "../schemas/validateDocument";
@@ -55,6 +46,8 @@ import { WebviewPanelCache } from "../webview-cache";
 import { handleWebviewMessage } from "../webview/comms/comms";
 import { post } from "../webview/topic-config-form";
 import topicFormTemplate from "../webview/topic-config-form.html";
+import { createProduceRequestData } from "./utils/produceMessage";
+import { ProduceMessageSchemaOptions } from "./utils/types";
 
 const logger = new Logger("topics");
 
@@ -354,13 +347,6 @@ export async function produceMessagesFromDocument(topic: KafkaTopic) {
   }
 }
 
-export interface ProduceMessageSchemaOptions {
-  keySchema?: Schema;
-  valueSchema?: Schema;
-  keySubjectNameStrategy?: SubjectNameStrategy;
-  valueSubjectNameStrategy?: SubjectNameStrategy;
-}
-
 /**
  * Produce multiple messages to a Kafka topic through a worker pool and display info and/or error
  * notifications depending on the {@link ExecutionResult}s.
@@ -527,70 +513,6 @@ export async function produceMessage(
   return { timestamp, response };
 }
 
-/** Create the {@link ProduceRequestData} objects for the `key` and `value` of a produce request based on the provided `message` content and any schema options. */
-export async function createProduceRequestData(
-  message: ProduceMessage,
-  schemaOptions: ProduceMessageSchemaOptions = {},
-  forCCloudTopic: boolean = false,
-): Promise<{ keyData: ProduceRequestData; valueData: ProduceRequestData }> {
-  // snake case since this is coming from a JSON document:
-  const { key, value, key_schema, value_schema } = message;
-  // user-selected schema information via settings and quickpicks
-  const { keySchema, keySubjectNameStrategy, valueSchema, valueSubjectNameStrategy } =
-    schemaOptions;
-
-  // determine if we have to provide `type` based on whether this is a CCloud-flavored topic or not
-  const schemaless = "JSON";
-  const schemaType: { type?: string } = {};
-  if (forCCloudTopic && !(keySchema || key_schema || valueSchema || value_schema)) {
-    schemaType.type = schemaless;
-  }
-
-  // message-provided schema information takes precedence over quickpicked schema
-  const keySchemaData: SchemaInfo | undefined = extractSchemaInfo(
-    key_schema,
-    keySchema,
-    keySubjectNameStrategy,
-  );
-  const keyData: ProduceRequestData = {
-    ...schemaType,
-    ...(keySchemaData ?? {}),
-    data: key,
-  };
-  const valueSchemaData: SchemaInfo | undefined = extractSchemaInfo(
-    value_schema,
-    valueSchema,
-    valueSubjectNameStrategy,
-  );
-  const valueData: ProduceRequestData = {
-    ...schemaType,
-    ...(valueSchemaData ?? {}),
-    data: value,
-  };
-  return { keyData, valueData };
-}
-
-/**
- * Extract schema information from provided produce-message content, or a {@link Schema}. Returns
- * the necessary {@link SchemaInfo} object for the produce request.
- */
-export function extractSchemaInfo(
-  schemaInfo: any,
-  schema: Schema | undefined,
-  subjectNameStrategy: SubjectNameStrategy | undefined,
-): SchemaInfo | undefined {
-  if (!(schemaInfo || schema)) {
-    return;
-  }
-  const schema_version = schemaInfo?.schema_version ?? schema?.version;
-  const subject = schemaInfo?.subject ?? schema?.subject;
-  const subject_name_strategy =
-    schemaInfo?.subject_name_strategy ?? subjectNameStrategy ?? "TOPIC_NAME";
-
-  // drop type since the sidecar rejects this with a 400
-  return { schema_version, subject, subject_name_strategy, type: undefined };
-}
-
 export function registerTopicCommands(): vscode.Disposable[] {
   return [
     registerCommandWithLogging("confluent.topics.copyKafkaClusterId", copyKafkaClusterId),
@@ -602,133 +524,4 @@ export function registerTopicCommands(): vscode.Disposable[] {
     registerCommandWithLogging("confluent.topics.edit", editTopicConfig),
     registerCommandWithLogging("confluent.topic.produce.fromDocument", produceMessagesFromDocument),
   ];
-}
-
-/**
- * Return {@linkcode SubjectNameStrategy.TOPIC_NAME} if enabled in user settings, otherwise prompt for
- * the user to select a subject name strategy via quickpick.
- */
-export async function getSubjectNameStrategy(
-  topic: KafkaTopic,
-  kind: "key" | "value",
-): Promise<SubjectNameStrategy | undefined> {
-  const config: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration();
-  const useTopicNameStrategy: boolean = config.get(USE_TOPIC_NAME_STRATEGY) ?? true;
-  if (useTopicNameStrategy) {
-    return SubjectNameStrategy.TOPIC_NAME;
-  }
-  // if the user has disabled the topic name strategy, we need to prompt for the subject name
-  // strategy first, which will help narrow down subjects later
-  return await subjectNameStrategyQuickPick(topic, kind);
-}
-
-/**
- * Prompt the user to select a schema subject+version to use when producing messages to a topic.
- * @param topic The Kafka topic to produce messages to
- * @param kind Whether this is for a 'key' or 'value' schema
- * @returns The selected {@link Schema}, or `undefined` if user cancelled
- */
-export async function promptForSchema(
-  topic: KafkaTopic,
-  kind: "key" | "value",
-  strategy: SubjectNameStrategy,
-): Promise<Schema> {
-  // look up the associated SR instance for this topic
-  const loader = ResourceLoader.getInstance(topic.connectionId);
-  const schemaRegistries: SchemaRegistry[] = await loader.getSchemaRegistries();
-  const registry: SchemaRegistry | undefined = schemaRegistries.find(
-    (registry) => registry.environmentId === topic.environmentId,
-  );
-  if (!registry) {
-    const noRegistryMsg = `No Schema Registry available for topic "${topic.name}".`;
-    showErrorNotificationWithButtons(noRegistryMsg);
-    throw new Error(noRegistryMsg);
-  }
-
-  let schemaSubject: string | undefined = await getSubjectNameForStrategy(
-    strategy,
-    topic,
-    kind,
-    registry,
-    loader,
-  );
-  if (!schemaSubject) {
-    throw new Error(`"${kind}" schema subject not found/set for topic "${topic.name}".`);
-  }
-
-  const config: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration();
-  const allowOlderSchemaVersions: boolean = config.get(ALLOW_OLDER_SCHEMA_VERSIONS, false);
-  if (allowOlderSchemaVersions) {
-    // allow the user to select a specific schema version if they enabled the setting
-    const schemaVersion: Schema | undefined = await schemaVersionQuickPick(registry, schemaSubject);
-    if (!schemaVersion) {
-      throw new Error("Schema version not chosen.");
-    }
-    return schemaVersion;
-  }
-
-  // look up the latest schema version for the given subject
-  const schemaVersions: Schema[] = await loader.getSchemasForEnvironmentId(registry.environmentId);
-  const latestSchema: Schema | undefined = schemaVersions
-    .filter((schema) => schema.subject === schemaSubject)
-    .sort((a, b) => b.version - a.version)[0];
-  if (!latestSchema) {
-    const noVersionsMsg = `No schema versions found for subject "${schemaSubject}".`;
-    showErrorNotificationWithButtons(noVersionsMsg);
-    throw new Error(noVersionsMsg);
-  }
-  return latestSchema;
-}
-
-async function getSubjectNameForStrategy(
-  strategy: SubjectNameStrategy,
-  topic: KafkaTopic,
-  kind: string,
-  registry: SchemaRegistry,
-  loader: ResourceLoader,
-): Promise<string | undefined> {
-  let schemaSubjectName: string | undefined;
-
-  switch (strategy) {
-    case SubjectNameStrategy.TOPIC_NAME:
-      {
-        // we have the topic name and the kind, so we just need to make sure the subject exists and
-        // fast-track to getting the schema version
-        schemaSubjectName = `${topic.name}-${kind}`;
-        const schemaSubjects: Subject[] = await loader.getSubjects(registry);
-        const subjectExists = schemaSubjects.some((s) => s.name === schemaSubjectName);
-        if (!subjectExists) {
-          const noSubjectMsg = `No "${kind}" schema subject found for topic "${topic.name}" using the ${strategy} strategy.`;
-          showErrorNotificationWithButtons(noSubjectMsg, {
-            "Open Settings": () => {
-              vscode.commands.executeCommand(
-                "workbench.action.openSettings",
-                `@id:confluent.topic.produceMessages.schemas.useTopicNameStrategy`,
-              );
-            },
-            ...DEFAULT_ERROR_NOTIFICATION_BUTTONS,
-          });
-          throw new Error(noSubjectMsg);
-        }
-      }
-      break;
-    case SubjectNameStrategy.TOPIC_RECORD_NAME:
-      // filter the subject quickpick based on the topic name
-      schemaSubjectName = await schemaSubjectQuickPick(
-        registry,
-        false,
-        `Producing to ${topic.name}: ${kind} schema`,
-        (s) => s.name.startsWith(topic.name),
-      );
-      break;
-    case SubjectNameStrategy.RECORD_NAME:
-      schemaSubjectName = await schemaSubjectQuickPick(
-        registry,
-        false,
-        `Producing to ${topic.name}: ${kind} schema`,
-      );
-      break;
-  }
-
-  return schemaSubjectName;
 }
