+++ src/quickpicks/utils/README.md
@@ -1,0 +1,5 @@
+This subdirectory is for utility functions used by the quickpick-functions themselves. Separating
+them from the quickpick modules will help readability/maintainability while also better supporting
+test suite setup.
+
+Any functions that are not specific to quickpicks should be moved into a higher-level module.
